// src/app/models/withdrawal/WithdrawalModal.ts
export class WithdrawalModal {
  internalReferenceNumber: number; 
  userId: string;
  userName: string;
  accountNumber: string;
  accountHolderName: string;
  ifscCode: string;
  paymentTransactionAmount: string;
  paymentMode: string;
  vpa: string;
  transactionStatusDescription: string;
  reason: string;
  tdsAmount: string;
  toBePaidAmount: string;
  withdrawalType: string;
  nameMismatch: string;
  isSuspicious: string;
  fulfilledTime: string;
  createdAt: string;
  updatedAt: string;
}